started refactoring
- made all evaluation "themes" into python packages
- created a script that can run any evaluation,
- and have an initial golds.yaml that I will update as I start to update goldretriever.

I had to make some small refactoring changes to some of the evaluate.py files of the subdirectories in order for them to be compatible, and added some test files and test gold files for testing the evaluator. The golds directory will eventually be combined with golds.yaml for simmplicity.